requirements.txt
Flask>=2.0.2
Flask-HTTPAuth>=4.7.0
RPi.GPIO>=0.7.0